Modification in libnegf_api.f90, routine set_s_id

+++ src/api/libnegf_api.f90
@@ -429,7 +429,7 @@ subroutine negf_set_s_id(handler, nrow) bind(C)
   use libnegf           ! if:mod:use
   use ln_precision      ! if:mod:use
   implicit none
-  integer(c_int) :: handler(DAC_handlerSize)  ! if:var:in
+  integer :: handler(DAC_handlerSize)  ! if:var:in
   integer(c_int), intent(in), value:: nrow     ! if:var:out
 
   type(NEGFpointers) :: LIB
